fix: worksapce sidebar item refactor
