+++ packages/i18n/src/locales/es/translations.json
@@ -314,5 +314,6 @@
   "change_parent_issue": "Cambiar problema padre",
   "remove_parent_issue": "Eliminar problema padre",
   "add_parent": "Agregar padre",
-  "loading_members": "Cargando miembros..."
+  "loading_members": "Cargando miembros...",
+  "inbox": "bandeja de entrada"
 }
