+++ packages/i18n/src/locales/ja/translations.json
@@ -314,5 +314,6 @@
   "change_parent_issue": "親問題を変更",
   "remove_parent_issue": "親問題を削除",
   "add_parent": "親問題を追加",
-  "loading_members": "メンバーを読み込んでいます..."
+  "loading_members": "メンバーを読み込んでいます...",
+  "inbox": "受信箱"
 }
