+++ web/ce/constants/dashboard.ts
@@ -25,49 +25,3 @@ export type TSidebarMenuItems<T extends TSidebarUserMenuItemKeys | TSidebarWorks
 export type TSidebarUserMenuItems = TSidebarMenuItems<TSidebarUserMenuItemKeys>;
 
 export type TSidebarWorkspaceMenuItems = TSidebarMenuItems<TSidebarWorkspaceMenuItemKeys>;
-
-export const SIDEBAR_WORKSPACE_MENU: Partial<Record<TSidebarWorkspaceMenuItemKeys, TSidebarWorkspaceMenuItems>> = {
-  projects: {
-    value: "projects",
-    key: "projects",
-    label: "Projects",
-    href: `/projects`,
-    access: [EUserPermissions.ADMIN, EUserPermissions.MEMBER, EUserPermissions.GUEST],
-    highlight: (pathname: string, baseUrl: string) => pathname === `${baseUrl}/projects/`,
-    Icon: Briefcase,
-  },
-  "all-issues": {
-    value: "all-issues",
-    key: "views",
-    label: "Views",
-    href: `/workspace-views/all-issues`,
-    access: [EUserPermissions.ADMIN, EUserPermissions.MEMBER, EUserPermissions.GUEST],
-    highlight: (pathname: string, baseUrl: string) => pathname.includes(`${baseUrl}/workspace-views/`),
-    Icon: Layers,
-  },
-  "active-cycles": {
-    value: "active-cycles",
-    key: "active_cycles",
-    label: "Cycles",
-    href: `/active-cycles`,
-    access: [EUserPermissions.ADMIN, EUserPermissions.MEMBER],
-    highlight: (pathname: string, baseUrl: string) => pathname === `${baseUrl}/active-cycles/`,
-    Icon: ContrastIcon,
-  },
-  analytics: {
-    value: "analytics",
-    key: "analytics",
-    label: "Analytics",
-    href: `/analytics`,
-    access: [EUserPermissions.ADMIN, EUserPermissions.MEMBER],
-    highlight: (pathname: string, baseUrl: string) => pathname.includes(`${baseUrl}/analytics/`),
-    Icon: BarChart2,
-  },
-};
-
-export const SIDEBAR_WORKSPACE_MENU_ITEMS: TSidebarWorkspaceMenuItems[] = [
-  SIDEBAR_WORKSPACE_MENU?.projects,
-  SIDEBAR_WORKSPACE_MENU?.["all-issues"],
-  SIDEBAR_WORKSPACE_MENU?.["active-cycles"],
-  SIDEBAR_WORKSPACE_MENU?.analytics,
-].filter((item): item is TSidebarWorkspaceMenuItems => item !== undefined);
