+++ web/core/components/workspace/sidebar/index.ts
@@ -8,3 +8,5 @@ export * from "./quick-actions";
 export * from "./user-menu";
 export * from "./user-menu-item";
 export * from "./workspace-menu";
+export * from "./workspace-menu-item";
+export * from "./workspace-menu-header";
