+++ web/core/components/workspace/sidebar/user-menu.tsx
@@ -24,28 +24,28 @@ export const SidebarUserMenu = observer(() => {
     {
       key: "home",
       labelTranslationKey: "home",
-      href: `/${workspaceSlug.toString()}`,
+      href: `/${workspaceSlug.toString()}/`,
       access: [EUserWorkspaceRoles.ADMIN, EUserWorkspaceRoles.MEMBER, EUserWorkspaceRoles.GUEST],
       Icon: Home,
     },
     {
       key: "your-work",
       labelTranslationKey: "your_work",
-      href: `/${workspaceSlug.toString()}/profile/${currentUser?.id}`,
+      href: `/${workspaceSlug.toString()}/profile/${currentUser?.id}/`,
       access: [EUserWorkspaceRoles.ADMIN, EUserWorkspaceRoles.MEMBER],
       Icon: UserActivityIcon,
     },
     {
       key: "notifications",
       labelTranslationKey: "inbox",
-      href: `/${workspaceSlug.toString()}/notifications`,
+      href: `/${workspaceSlug.toString()}/notifications/`,
       access: [EUserWorkspaceRoles.ADMIN, EUserWorkspaceRoles.MEMBER, EUserWorkspaceRoles.GUEST],
       Icon: Inbox,
     },
     {
       key: "drafts",
       labelTranslationKey: "drafts",
-      href: `/${workspaceSlug.toString()}/drafts`,
+      href: `/${workspaceSlug.toString()}/drafts/`,
       access: [EUserWorkspaceRoles.ADMIN, EUserWorkspaceRoles.MEMBER],
       Icon: PenSquare,
     },
